+++ xlive/Blam/Cache/TagGroups/scenario_structure_bsp_definition.hpp
@@ -3,7 +3,7 @@
 #include "Blam/Cache/DataTypes/BlamDataTypes.h"
 #include "Blam/Cache/DataTypes/BlamPrimitiveType.h"
 #include "Blam/Cache/TagGroups.hpp"
-#include "Blam/Math/real_math.h"
+#include "Blam/Engine/math/real_math.h"
 
 /*********************************************************************
 * name: scenario_structure_bsp
