+++ xlive/Blam/Cache/TagGroups/sound_classes_definition.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "Blam/Cache/DataTypes/BlamDataTypes.h"
 #include "Blam/Cache/TagGroups.hpp"
-#include "Blam/Math/real_math.h"
+#include "Blam/Engine/math/real_math.h"
 #include <wtypes.h>
 
 #pragma pack(push,1)
