+++ xlive/Blam/Engine/physics/bsp3d.h
@@ -1,5 +1,5 @@
 #pragma once
 #include "Blam/Cache/TagGroups/scenario_structure_bsp_definition.hpp"
-#include "Blam/Math/real_math.h"
+#include "Blam/Engine/math/real_math.h"
 
 int bsp3d_test_point(const s_scenario_structure_bsp_group_definition::s_collision_bsp_block* collision_bsp, int starting_index, const real_point3d* point);
