+++ xlive/Blam/Engine/scenario/scenario.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "Blam/Cache/DataTypes/BlamDataTypes.h"
 #include "Blam/Engine/structures/structures.h"
-#include "Blam/Math/real_math.h"
+#include "Blam/Engine/math/real_math.h"
 
 enum class e_placement_flags : __int32
 {
